componetizando a cesta
