+++ App.js
@@ -1,9 +1,11 @@
 
 import React from 'react';
 import { StatusBar, SafeAreaView, View } from 'react-native';
-import { useFonts,
+import {
+  useFonts,
   Montserrat_400Regular,
-  Montserrat_700Bold }
+  Montserrat_700Bold
+}
   from '@expo-google-fonts/montserrat';
 
 import Cesta from './src/telas/Cesta';
@@ -14,14 +16,14 @@ export default function App() {
     "MontserratBold": Montserrat_700Bold
   });
 
-  if(!fontCarregada){
+  if (!fontCarregada) {
     return <View />
   }
 
   return (
     <SafeAreaView>
       <StatusBar />
-      <Cesta/>
+      <Cesta />
     </SafeAreaView>
   );
 }
