+++ src/telas/Cesta/componentes/Topo.js
@@ -1,0 +1,31 @@
+import React from "react";
+import { Image, StyleSheet, Dimensions } from "react-native";
+import Texto from "../../../componentes/Texto";
+import topo from '../../../../assets/topo.png';
+
+const width = Dimensions.get('screen').width;
+
+export default function Topo() {
+    return <>
+        <Image source={topo} style={estilos.topo} />
+        <Texto style={estilos.titulo}>Detalhe da Cesta</Texto>
+    </>
+}
+
+const estilos = StyleSheet.create({
+    topo: {
+        width: "100%",
+        height: 578 / 768 * width,
+        fontFamily: "Montserrat"
+    },
+    titulo: {
+        width: "100%",
+        position: "absolute",
+        textAlign: "center",
+        fontSize: 16,
+        lineHeight: 26,
+        color: "white",
+        fontWeight: "bold",
+        padding: 16
+    }
+})
